+++ satella/json.py
@@ -32,8 +32,15 @@ class JSONEncoder(json.JSONEncoder):
             return super().default(o)
         except TypeError:
             dct = {}
-            for k, v in o.__dict__.items():
-                dct[k] = repr(v)
+            try:
+                for k, v in o.__dict__.items():
+                    dct[k] = repr(v)
+            except AttributeError:  # o has no attribute '__dict__', try with slots
+                try:
+                    for slot in o.__slots__:
+                        dct[slot] = getattr(o, slot)
+                except AttributeError:      # it doesn't have __slots__ either?
+                    return '<an instance of %s>' % (o.__class__.__name__, )
             return dct