+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * added `dont_raise` to `Condition`
 * added `ImpossibleError`
 * added `Predicate.false`
+* fixed behaviour of `JSONEncoder` when faced with objects without __dict__ and __slots__